requirements.txt
irc
ultraimport
discord.py
dotenv
